# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NWB Vending Materials Factory",
    "version": "1.3.0",
    "description": "x402 agent inventory: training, tools, and template packs for small-business ops.",
    "contact": {
      "name": "NWB Vending",
      "email": "NWBVending@gmail.com"
    },
    "x-guidance": "Discover packs via catalog-manifest.json. Each GET path returns a paid JSON artifact after x402 USDC payment on Base. Start with catalog-manifest.json (public) for routing; fetch primary JSON files listed here when use_when matches."
  },
  "servers": [
    {
      "url": "https://nwb-vending.com"
    }
  ],
...
  "paths": {
    "/2nd-pages/materials-factory/training/llm-api-bootcamp/full/playbook.json": {
      "get": {
        "operationId": "mf-train-001",
        "summary": "LLM API Bootcamp",
        "description": "Materials Factory training pack mf-train-001. Paid JSON for agents.",
        "tags": [
          "training"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.08"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "pack_id",
            "in": "query",
            "required": false,
            "description": "Catalog pack id (informational). Primary artifact is returned from the path after x402 payment.",
            "schema": {
              "type": "string",
              "enum": [
                "mf-train-001"
              ],
              "default": "mf-train-001"
            },
            "example": "mf-train-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid agent pack JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "audience": {
                      "type": "string",
                      "enum": [
                        "agent"
                      ]
                    }
                  },
                  "required": [
                    "artifact",
                    "title",
                    "audience"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/2nd-pages/materials-factory/tools/should-i-automate/full/rules.json": {
      "get": {
        "operationId": "mf-tool-001",
        "summary": "Should I Automate This?",
        "description": "Deterministic rubric: should this task stay manual, use a template, gather data, or automate? Built for ops-heavy small business.",
        "tags": [
          "tools"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.05"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "pack_id",
            "in": "query",
            "required": false,
            "description": "Catalog pack id (informational). Primary artifact is returned from the path after x402 payment.",
            "schema": {
              "type": "string",
              "enum": [
                "mf-tool-001"
              ],
              "default": "mf-tool-001"
            },
            "example": "mf-tool-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid agent pack JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "audience": {
                      "type": "string",
                      "enum": [
                        "agent"
                      ]
                    }
                  },
                  "required": [
                    "artifact",
                    "title",
                    "audience"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/2nd-pages/materials-factory/tools/ai-task-fit/full/rules.json": {
      "get": {
        "operationId": "mf-tool-002",
        "summary": "AI Task Fit",
        "description": "Route tasks to AI draft, human-only, or template path. Pairs with job spine comms packs.",
        "tags": [
          "tools"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.05"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "pack_id",
            "in": "query",
            "required": false,
            "description": "Catalog pack id (informational). Primary artifact is returned from the path after x402 payment.",
            "schema": {
              "type": "string",
              "enum": [
                "mf-tool-002"
              ],
              "default": "mf-tool-002"
            },
            "example": "mf-tool-002"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid agent pack JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "audience": {
                      "type": "string",
                      "enum": [
                        "agent"
                      ]
                    }
                  },
                  "required": [
                    "artifact",
                    "title",
                    "audience"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/2nd-pages/materials-factory/tools/scope-creep-say-no/full/rules.json": {
      "get": {
        "operationId": "mf-tool-003",
        "summary": "Scope Creep / Say-No",
        "description": "Say-no and CO routing rubric for scope creep. Cross-links mf-tpl-012 and quote gate.",
        "tags": [
          "tools"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.05"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "pack_id",
            "in": "query",
            "required": false,
            "description": "Catalog pack id (informational). Primary artifact is returned from the path after x402 payment.",
            "schema": {
              "type": "string",
              "enum": [
                "mf-tool-003"
              ],
              "default": "mf-tool-003"
            },
            "example": "mf-tool-003"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid agent pack JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "audience": {
                      "type": "string",
                      "enum": [
                        "agent"
                      ]
                    }
                  },
                  "required": [
                    "artifact",
                    "title",
                    "audience"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/2nd-pages/materials-factory/templates/complaint-recovery-playbook/full/playbook.json": {
      "get": {
        "operationId": "mf-tpl-001",
        "summary": "Customer Complaint Recovery",
        "description": "Complaint recovery spine route: severity branches, templates, log rows. Use before treating anger as new intake.",
        "tags": [
          "templates"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.08"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "pack_id",
            "in": "query",
            "required": false,
            "description": "Catalog pack id (informational). Primary artifact is returned from the path after x402 payment.",
            "schema": {
              "type": "string",
              "enum": [
                "mf-tpl-001"
              ],
              "default": "mf-tpl-001"
            },
            "example": "mf-tpl-001"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid agent pack JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "audience": {
                      "type": "string",
                      "enum": [
                        "agent"
                      ]
                    }
                  },
                  "required": [
                    "artifact",
                    "title",
                    "audience"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/2nd-pages/materials-factory/templates/employee-onboarding-kit/full/playbook.json": {
      "get": {
        "operationId": "mf-tpl-002",
        "summary": "Rapid Employee Onboarding",
        "description": "Rapid onboarding by role: templates, buddy, Day 7/30 checkpoints. HR adjacent to job spine.",
        "tags": [
          "templates"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.08"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "pack_id",
            "in": "query",
            "required": false,
            "description": "Catalog pack id (informational). Primary artifact is returned from the path after x402 payment.",
            "schema": {
              "type": "string",
              "enum": [
                "mf-tpl-002"
              ],
              "default": "mf-tpl-002"
            },
            "example": "mf-tpl-002"
          }
        ],
        "responses": {
          "200": {
            "description": "Paid agent pack JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "artifact": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    },
                    "audience": {
                      "type": "string",
                      "enum": [
                        "agent"
                      ]
                    }
                  },
                  "required": [
                    "artifact",
                    "title",
                    "audience"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
...
  }
}